requirements.txt
asgiref==3.4.1
Django==3.2.7
pip==21.2.4
pytz==2021.1
setuptools==58.1.0
sqlparse==0.4.2
wheel==0.37.0

djangorestframework~=3.12.4